PYL-26 finalize refactor and move cli related to module cli
